+++ adapters/LiftoffMonetize/LiftoffMonetizeAdapter/Bidding/GADMediationVungleRewardedAd.h
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,16 +15,16 @@
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface GADMediationVungleInterstitial : NSObject
+@interface GADMediationVungleRewardedAd : NSObject <GADMediationRewardedAd>
 
 - (nonnull instancetype)
-    initWithAdConfiguration:(nonnull GADMediationInterstitialAdConfiguration *)adConfiguration
-          completionHandler:(nonnull GADMediationInterstitialLoadCompletionHandler)handler;
+    initWithAdConfiguration:(nonnull GADMediationRewardedAdConfiguration *)adConfiguration
+          completionHandler:(nonnull GADMediationRewardedLoadCompletionHandler)handler;
 
 /// Constructor is unavailable. Please use initWithAdConfiguration:completionHandler:.
 - (nonnull instancetype)init NS_UNAVAILABLE;
 
-/// Requests a interstitial ad from Vungle.
-- (void)requestInterstitialAd;
+/// Requests a rewarded ad from Liftoff Monetize.
+- (void)requestRewardedAd;
 
 @end